[VSC-6] Untie handlers from connection
